Big giant warnings cleanup, thanks to Eclipse.
Fix all raw type warnings, attaching suitable type arguments to all
generic class usages. Change all "bag" occurrences to use Set instead.
Use autoboxing/autounboxing where appropriate.

Remove the ArraySorter, Comparer, and VectorSorter. These are completely
redundant given the existence of Arrays.sort and Comparable.

+++ src/com/sun/tools/hat/internal/model/JavaByte.java
@@ -48,7 +48,7 @@ public class JavaByte extends JavaValue {
     }
 
     public String toString() {
-        return "0x" + Integer.toString(((int) value) & 0xff, 16);
+        return "0x" + Integer.toString(value & 0xff, 16);
     }
 
 }
